requirements.txt
six
future
nibabel
numpy
scipy
pydicom >= 1.0.1